kiosk: title-based config, auto-save, focus-mode chrome, polish
Refines the kiosk feature based on review:

- Search and store by Title (not Model). The kiosk's user-visible unit
  matches how museum visitors think — "Godzilla" — and the configure
  search now hits /api/titles/all/. KioskItem.modelSlug -> titleSlug.
  Visitor cards link to /titles/{slug}, the canonical detail page that
  surfaces all variants.
- Reuse existing matchesQuery from facet-engine.ts (now exported with a
  structural param type) instead of writing a parallel matcher.
- Hide nav/footer on /kiosk and /kiosk/configure by adding them to
  isFocusModePath.
- Auto-save on every edit via $effect; remove Save and Launch buttons.
  Configure page now has a single Enter/Exit Kiosk Mode toggle button
  (using the project's Button component).
- Visitor grid extracted to $lib/kiosk/KioskHome.svelte for reuse.
- Idle-reset exempts /kiosk/configure so staff edits aren't blown away.
- Card meta separator fix; small CSS cleanup.
- Move "Add a machine" search above the Machines (N) heading; the
  heading now scopes only the configured list.
