+++ frontend/src/routes/kiosk/configure/+page.svelte
@@ -1,7 +1,6 @@
 <!--
   Kiosk configuration page. Staff use this to choose machines, edit hooks,
-  reorder, and launch the kiosk. Configuration is saved to localStorage on
-  the device.
+  reorder, and toggle kiosk mode. Every edit auto-saves to localStorage.
 -->
 <script lang="ts">
   import { onMount } from 'svelte';
@@ -20,28 +19,30 @@
     type KioskItem,
   } from '$lib/kiosk/config';
   import { normalizeText } from '$lib/utils';
-  import type { ModelGridItemSchema } from '$lib/api/schema';
+  import { matchesQuery } from '$lib/facet-engine';
+  import Button from '$lib/components/Button.svelte';
+  import type { TitleListItemSchema } from '$lib/api/schema';
 
   let title = $state(DEFAULT_TITLE);
   let idleSeconds = $state(DEFAULT_IDLE_SECONDS);
   let items = $state<KioskItem[]>([]);
-  let allModels = $state<ModelGridItemSchema[]>([]);
+  let allTitles = $state<TitleListItemSchema[]>([]);
   let search = $state('');
   let kioskActive = $state(false);
-  let saved = $state(false);
+  let loaded = $state(false);
 
-  let configuredSlugs = $derived(new Set(items.map((i) => i.modelSlug)));
+  let configuredSlugs = $derived(new Set(items.map((i) => i.titleSlug)));
 
   let searchResults = $derived.by(() => {
     const q = normalizeText(search);
     if (!q) return [];
-    return allModels
-      .filter((m) => !configuredSlugs.has(m.slug))
-      .filter((m) => normalizeText(m.search_text ?? m.name).includes(q))
+    return allTitles
+      .filter((t) => !configuredSlugs.has(t.slug))
+      .filter((t) => matchesQuery(t, q))
       .slice(0, 12);
   });
 
-  let modelBySlug = $derived(new Map(allModels.map((m) => [m.slug, m])));
+  let titleBySlug = $derived(new Map(allTitles.map((t) => [t.slug, t])));
 
   onMount(async () => {
     const existing = loadConfig();
@@ -51,19 +52,36 @@
       items = [...existing.items];
     }
     kioskActive = isKioskCookieSet();
-    const res = await client.GET('/api/models/all/');
-    if (res.data) allModels = res.data;
+    loaded = true;
+    const res = await client.GET('/api/titles/all/');
+    if (res.data) allTitles = res.data;
   });
 
-  function addModel(slug: string) {
-    items = [...items, { modelSlug: slug, hook: '' }];
+  function buildConfig(): KioskConfig {
+    return {
+      title: title.trim() || DEFAULT_TITLE,
+      idleSeconds: idleSeconds > 0 ? idleSeconds : DEFAULT_IDLE_SECONDS,
+      items: items.map((i) => ({
+        titleSlug: i.titleSlug,
+        hook: i.hook.slice(0, HOOK_MAX_LENGTH),
+      })),
+    };
+  }
+
+  // Auto-save: any edit to title, idleSeconds, or items writes to localStorage.
+  // The `loaded` gate prevents writes during the initial onMount setters.
+  $effect(() => {
+    if (!loaded) return;
+    saveConfig(buildConfig());
+  });
+
+  function addTitle(slug: string) {
+    items = [...items, { titleSlug: slug, hook: '' }];
     search = '';
-    saved = false;
   }
 
   function removeAt(index: number) {
     items = items.filter((_, i) => i !== index);
-    saved = false;
   }
 
   function moveUp(index: number) {
@@ -71,7 +89,6 @@
     const next = [...items];
     [next[index - 1], next[index]] = [next[index], next[index - 1]];
     items = next;
-    saved = false;
   }
 
   function moveDown(index: number) {
@@ -79,27 +96,9 @@
     const next = [...items];
     [next[index + 1], next[index]] = [next[index], next[index + 1]];
     items = next;
-    saved = false;
   }
 
-  function buildConfig(): KioskConfig {
-    return {
-      title: title.trim() || DEFAULT_TITLE,
-      idleSeconds: idleSeconds > 0 ? idleSeconds : DEFAULT_IDLE_SECONDS,
-      items: items.map((i) => ({
-        modelSlug: i.modelSlug,
-        hook: i.hook.slice(0, HOOK_MAX_LENGTH),
-      })),
-    };
-  }
-
-  function handleSave() {
-    saveConfig(buildConfig());
-    saved = true;
-  }
-
-  async function handleLaunch() {
-    saveConfig(buildConfig());
+  async function handleEnter() {
     setKioskCookie();
     await goto('/kiosk');
   }
@@ -118,7 +117,9 @@
   <header class="header">
     <h1>Configure Kiosk</h1>
     {#if kioskActive}
-      <button type="button" class="exit-btn" onclick={handleExit}>Exit kiosk mode</button>
+      <Button type="button" onclick={handleExit}>Exit Kiosk Mode</Button>
+    {:else}
+      <Button type="button" onclick={handleEnter}>Enter Kiosk Mode</Button>
     {/if}
   </header>
 
@@ -134,8 +135,6 @@
   </section>
 
   <section class="machines">
-    <h2>Machines ({items.length})</h2>
-
     <label class="search">
       <span>Add a machine</span>
       <input type="search" placeholder="Search by name…" bind:value={search} autocomplete="off" />
@@ -143,13 +142,13 @@
 
     {#if searchResults.length > 0}
       <ul class="results">
-        {#each searchResults as model (model.slug)}
+        {#each searchResults as t (t.slug)}
           <li>
-            <button type="button" onclick={() => addModel(model.slug)}>
-              <strong>{model.name}</strong>
+            <button type="button" onclick={() => addTitle(t.slug)}>
+              <strong>{t.name}</strong>
               <span class="result-meta">
-                {#if model.manufacturer_name}{model.manufacturer_name}{/if}
-                {#if model.year}· {model.year}{/if}
+                {#if t.manufacturer}{t.manufacturer.name}{/if}
+                {#if t.year}· {t.year}{/if}
               </span>
             </button>
           </li>
@@ -157,20 +156,20 @@
       </ul>
     {/if}
 
+    <h2>Machines ({items.length})</h2>
+
     {#if items.length === 0}
       <p class="empty">No machines added yet.</p>
     {:else}
       <ol class="items">
-        {#each items as item, i (item.modelSlug)}
-          {@const model = modelBySlug.get(item.modelSlug)}
+        {#each items as item, i (item.titleSlug)}
+          {@const t = titleBySlug.get(item.titleSlug)}
           <li class="item">
             <div class="item-header">
               <span class="item-name">
-                {model?.name ?? item.modelSlug}
-                {#if model?.year}<span class="dim"> · {model.year}</span>{/if}
-                {#if model?.manufacturer_name}<span class="dim">
-                    · {model.manufacturer_name}</span
-                  >{/if}
+                {t?.name ?? item.titleSlug}
+                {#if t?.year}<span class="dim"> · {t.year}</span>{/if}
+                {#if t?.manufacturer}<span class="dim"> · {t.manufacturer.name}</span>{/if}
               </span>
               <div class="item-actions">
                 <button
@@ -193,19 +192,12 @@
               placeholder="Hook (optional, e.g. 'First talking pinball machine')"
               bind:value={item.hook}
               maxlength={HOOK_MAX_LENGTH}
-              oninput={() => (saved = false)}
             />
           </li>
         {/each}
       </ol>
     {/if}
   </section>
-
-  <footer class="actions">
-    <button type="button" onclick={handleSave}>Save</button>
-    <button type="button" class="primary" onclick={handleLaunch}>Launch kiosk</button>
-    {#if saved}<span class="saved-hint">Saved.</span>{/if}
-  </footer>
 </div>
 
 <style>
@@ -226,14 +218,6 @@
 
   .header h1 {
     margin: 0;
-  }
-
-  .exit-btn {
-    background: transparent;
-    border: 1px solid var(--color-border-soft);
-    padding: var(--size-2) var(--size-3);
-    border-radius: var(--radius-2);
-    cursor: pointer;
   }
 
   .settings {
@@ -263,7 +247,7 @@
   }
 
   .machines h2 {
-    margin: 0 0 var(--size-3);
+    margin: var(--size-5) 0 var(--size-3);
   }
 
   .results {
@@ -362,34 +346,5 @@
     color: var(--color-text-muted);
     font-style: italic;
     margin-top: var(--size-3);
-  }
-
-  .actions {
-    display: flex;
-    align-items: center;
-    gap: var(--size-3);
-    padding-top: var(--size-4);
-    border-top: 1px solid var(--color-border-soft);
-  }
-
-  .actions button {
-    padding: var(--size-2) var(--size-4);
-    border-radius: var(--radius-2);
-    border: 1px solid var(--color-border-soft);
-    background: transparent;
-    font-size: var(--font-size-1);
-    cursor: pointer;
-  }
-
-  .actions button.primary {
-    background: var(--color-accent);
-    color: white;
-    border-color: var(--color-accent);
-    font-weight: 600;
-  }
-
-  .saved-hint {
-    color: var(--color-text-muted);
-    font-size: var(--font-size-0);
   }
 </style>
